Update rust version and X86 crate
The x86 crate forces us to use u64 instead of usize in many places.
This is probably a good thing, although I'm not 100% sure I like that we
now use u64 instead of usize internally in the mem module for sizes.
We'll see how that works in the long run. Everywhere else should still
use usize.
The rust update changes how `static mut` referesnces are accessed. This
is now done using the `addr_of` of `addr_of_mut` macro.

+++ wasabi-kernel/src/graphics/fb.rs
@@ -62,7 +62,7 @@ pub struct Framebuffer {
 impl Framebuffer {
     /// Allocates a new memory backed framebuffer
     pub fn alloc_new(info: FrameBufferInfo) -> Result<Self, MemError> {
-        let page_count = (info.byte_len + Size4KiB::SIZE as usize - 1) / Size4KiB::SIZE as usize;
+        let page_count = (info.byte_len as u64 + Size4KiB::SIZE - 1) / Size4KiB::SIZE;
 
         let pages = PageAllocator::get_kernel_allocator()
             .lock()
